+++ changelog.md
@@ -15,6 +15,7 @@ While all _Major_ releases should be reviewed, our only _large_ releases are **v
 - **[change]** - tag text by default on .concat('')
 - **[change]** - allow changing term prePunctuation
 - **[new]** - .wrap() method
+- **[new]** - .isFull() method
 -->
 
 #### 14.6.0 [Oct 2022]